+++ plugins/clipboard/lang/ar.js
@@ -10,6 +10,5 @@ CKEDITOR.plugins.setLang( 'clipboard', 'ar', {
 	paste: 'لصق',
 	pasteNotification: 'Press %1 to paste. Your browser doesn‘t support pasting with the toolbar button or context menu option.', // MISSING
 	pasteArea: 'منطقة اللصق',
-	pasteMsg: 'Paste your content inside the area below and press OK.', // MISSING
-	title: 'لصق'
+	pasteMsg: 'Paste your content inside the area below and press OK.' // MISSING
 } );