+++ plugins/clipboard/lang/ca.js
@@ -10,6 +10,5 @@ CKEDITOR.plugins.setLang( 'clipboard', 'ca', {
 	paste: 'Enganxar',
 	pasteNotification: 'Press %1 to paste. Your browser doesn‘t support pasting with the toolbar button or context menu option.', // MISSING
 	pasteArea: 'Àrea d\'enganxat',
-	pasteMsg: 'Paste your content inside the area below and press OK.', // MISSING
-	title: 'Enganxar'
+	pasteMsg: 'Paste your content inside the area below and press OK.' // MISSING
 } );