+++ plugins/clipboard/lang/sl.js
@@ -10,6 +10,5 @@ CKEDITOR.plugins.setLang( 'clipboard', 'sl', {
 	paste: 'Prilepi',
 	pasteNotification: 'Press %1 to paste. Your browser doesn‘t support pasting with the toolbar button or context menu option.', // MISSING
 	pasteArea: 'Prilepi območje',
-	pasteMsg: 'Paste your content inside the area below and press OK.', // MISSING
-	title: 'Prilepi'
+	pasteMsg: 'Paste your content inside the area below and press OK.' // MISSING
 } );