+++ plugins/clipboard/lang/uk.js
@@ -10,6 +10,5 @@ CKEDITOR.plugins.setLang( 'clipboard', 'uk', {
 	paste: 'Вставити',
 	pasteNotification: 'Press %1 to paste. Your browser doesn‘t support pasting with the toolbar button or context menu option.', // MISSING
 	pasteArea: 'Область вставки',
-	pasteMsg: 'Paste your content inside the area below and press OK.', // MISSING
-	title: 'Вставити'
+	pasteMsg: 'Paste your content inside the area below and press OK.' // MISSING
 } );